+++ activerecord/test/cases/associations/has_one_through_associations_test.rb
@@ -15,6 +15,10 @@ require 'models/essay'
 require 'models/owner'
 require 'models/post'
 require 'models/comment'
+require 'models/customer'
+require 'models/carrier'
+require 'models/shop_account'
+require 'models/customer_carrier'
 
 class HasOneThroughAssociationsTest < ActiveRecord::TestCase
   fixtures :member_types, :members, :clubs, :memberships, :sponsors, :organizations, :minivans,
@@ -342,5 +346,35 @@ class HasOneThroughAssociationsTest < ActiveRecord::TestCase
         has_one :thing, through: :other_thing, counter_cache: true
       end
     end
+  end
+
+  def test_has_one_through_do_not_cache_association_reader_if_the_though_method_has_default_scopes
+    customer = Customer.create!
+    carrier = Carrier.create!
+    customer_carrier = CustomerCarrier.create!(
+      customer: customer,
+      carrier: carrier,
+    )
+    account = ShopAccount.create!(customer_carrier: customer_carrier)
+
+    CustomerCarrier.current_customer = customer
+
+    account_carrier = account.carrier
+    assert_equal carrier, account_carrier
+
+    CustomerCarrier.current_customer = nil
+
+    other_carrier = Carrier.create!
+    other_customer = Customer.create!
+    other_customer_carrier = CustomerCarrier.create!(
+      customer: other_customer,
+      carrier: other_carrier,
+    )
+    other_account = ShopAccount.create!(customer_carrier: other_customer_carrier)
+
+    account_carrier = other_account.carrier
+    assert_equal other_carrier, account_carrier
+  ensure
+    CustomerCarrier.current_customer = nil
   end
 end